Add ./pkg/util/cleaner
This is based on existing code in ./pkg/util/log
