+++ pkg/util/cleaner/cleaner_test.go
@@ -1,0 +1,85 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+package cleaner
+
+import (
+	"io/ioutil"
+	"path/filepath"
+	"regexp"
+	"strings"
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestRepl(t *testing.T) {
+	cleaner := New()
+	cleaner.AddReplacer(SingleLine, Replacer{
+		Regex: regexp.MustCompile("foo"),
+		Repl:  []byte("bar"),
+	})
+	res, err := cleaner.CredentialsCleanerBytes([]byte("dog food"))
+	require.NoError(t, err)
+	require.Equal(t, "dog bard", string(res))
+}
+
+func TestReplFunc(t *testing.T) {
+	cleaner := New()
+	cleaner.AddReplacer(SingleLine, Replacer{
+		Regex: regexp.MustCompile("foo"),
+		ReplFunc: func(match []byte) []byte {
+			return []byte(strings.ToUpper(string(match)))
+		},
+	})
+	res, err := cleaner.CredentialsCleanerBytes([]byte("dog food"))
+	require.NoError(t, err)
+	require.Equal(t, "dog FOOd", string(res))
+}
+
+func TestSkipCommentsAndBlanks(t *testing.T) {
+	cleaner := New()
+	cleaner.AddReplacer(SingleLine, Replacer{
+		Regex: regexp.MustCompile("foo"),
+		Repl:  []byte("bar"),
+	})
+	cleaner.AddReplacer(MultiLine, Replacer{
+		Regex: regexp.MustCompile("with bar\nanother"),
+		Repl:  []byte("..."),
+	})
+	res, err := cleaner.CredentialsCleanerBytes([]byte("a line with foo\n\n  \n  # a comment with foo\nanother line"))
+	require.NoError(t, err)
+	require.Equal(t, "a line ... line", string(res))
+}
+
+func TestCleanFile(t *testing.T) {
+	dir := t.TempDir()
+	filename := filepath.Join(dir, "test.yml")
+	ioutil.WriteFile(filename, []byte("a line with foo\n\na line with bar"), 0666)
+
+	cleaner := New()
+	cleaner.AddReplacer(SingleLine, Replacer{
+		Regex: regexp.MustCompile("foo"),
+		Repl:  []byte("bar"),
+	})
+	res, err := cleaner.CredentialsCleanerFile(filename)
+	require.NoError(t, err)
+	require.Equal(t, "a line with bar\na line with bar", string(res))
+}
+
+func TestSanitizeURL(t *testing.T) {
+	cleaner := New()
+	cleaner.AddReplacer(SingleLine, Replacer{
+		Regex: regexp.MustCompile(`([A-Za-z][A-Za-z0-9+-.]+\:\/\/|\b)([^\:]+)\:([^\s]+)\@`),
+		Repl:  []byte(`$1$2:********@`),
+	})
+	// this replacer should not be used on URLs!
+	cleaner.AddReplacer(MultiLine, Replacer{
+		Regex: regexp.MustCompile(".*"),
+		Repl:  []byte("UHOH"),
+	})
+	res := cleaner.SanitizeURL("https://foo:bar@example.com")
+	require.Equal(t, "https://foo:********@example.com", string(res))
+}
